+++ cognite/client/_api/data_modeling/containers.py
@@ -415,7 +415,7 @@ class ContainersAPI(APIClient):
                 ...     "prop21": ContainerPropertyApply(Json(is_list=True)),
                 ...     "prop22": ContainerPropertyApply(SequenceReference),
                 ...     "prop23": ContainerPropertyApply(SequenceReference(is_list=True)),
-                ...     # Note: It is adviced to represent files and time series directly as nodes
+                ...     # Note: It is advised to represent files and time series directly as nodes
                 ...     #       instead of referencing existing:
                 ...     "prop24": ContainerPropertyApply(FileReference),
                 ...     "prop25": ContainerPropertyApply(FileReference(is_list=True)),