+++ cognite/client/_sync_api/data_modeling/__init__.py
@@ -1,6 +1,6 @@
 """
 ===============================================================================
-8bb7b2ed7c8f40de00d57cc0eb15a566
+048014222247324af5e2bdcfee2e1b58
 This file is auto-generated from the Async API modules, - do not edit manually!
 ===============================================================================
 """
@@ -15,6 +15,7 @@ from cognite.client._sync_api.data_modeling.data_models import SyncDataModelsAPI
 from cognite.client._sync_api.data_modeling.files import SyncDataModelingFilesAPI
 from cognite.client._sync_api.data_modeling.graphql import SyncDataModelingGraphQLAPI
 from cognite.client._sync_api.data_modeling.instances import SyncInstancesAPI
+from cognite.client._sync_api.data_modeling.records import SyncRecordsAPI
 from cognite.client._sync_api.data_modeling.spaces import SyncSpacesAPI
 from cognite.client._sync_api.data_modeling.statistics import SyncStatisticsAPI
 from cognite.client._sync_api.data_modeling.streams import SyncStreamsAPI
@@ -37,5 +38,6 @@ class SyncDataModelingAPI(SyncAPIClient):
         self.views = SyncViewsAPI(async_client)
         self.instances = SyncInstancesAPI(async_client)
         self.graphql = SyncDataModelingGraphQLAPI(async_client)
+        self.records = SyncRecordsAPI(async_client)
         self.statistics = SyncStatisticsAPI(async_client)
         self.streams = SyncStreamsAPI(async_client)