+++ cognite/client/_sync_api/units.py
@@ -1,6 +1,6 @@
 """
 ===============================================================================
-7f8fbb89585e5600a0adc8cbcdab740b
+8fc99a6b213a3fc048bb5332fa2d8e84
 This file is auto-generated from the Async API modules, - do not edit manually!
 ===============================================================================
 """
@@ -48,14 +48,14 @@ class SyncUnitAPI(SyncAPIClient):
 
         Examples:
 
-            Retrive unit 'temperature:deg_c':
+            Retrieve unit 'temperature:deg_c':
 
                 >>> from cognite.client import CogniteClient, AsyncCogniteClient
                 >>> client = CogniteClient()
                 >>> # async_client = AsyncCogniteClient()  # another option
                 >>> res = client.units.retrieve("temperature:deg_c")
 
-            Retrive units 'temperature:deg_c' and 'pressure:bar':
+            Retrieve units 'temperature:deg_c' and 'pressure:bar':
 
                 >>> res = client.units.retrieve(["temperature:deg_c", "pressure:bar"])
         """