+++ cognite/client/testing.py
@@ -18,6 +18,7 @@ from cognite.client._api.data_modeling.data_models import DataModelsAPI
 from cognite.client._api.data_modeling.files import DataModelingFilesAPI
 from cognite.client._api.data_modeling.graphql import DataModelingGraphQLAPI
 from cognite.client._api.data_modeling.instances import InstancesAPI
+from cognite.client._api.data_modeling.records import RecordsAPI
 from cognite.client._api.data_modeling.space_statistics import SpaceStatisticsAPI
 from cognite.client._api.data_modeling.spaces import SpacesAPI
 from cognite.client._api.data_modeling.statistics import StatisticsAPI
@@ -103,6 +104,7 @@ from cognite.client._sync_api.data_modeling.data_models import SyncDataModelsAPI
 from cognite.client._sync_api.data_modeling.files import SyncDataModelingFilesAPI
 from cognite.client._sync_api.data_modeling.graphql import SyncDataModelingGraphQLAPI
 from cognite.client._sync_api.data_modeling.instances import SyncInstancesAPI
+from cognite.client._sync_api.data_modeling.records import SyncRecordsAPI
 from cognite.client._sync_api.data_modeling.space_statistics import SyncSpaceStatisticsAPI
 from cognite.client._sync_api.data_modeling.spaces import SyncSpacesAPI
 from cognite.client._sync_api.data_modeling.statistics import SyncStatisticsAPI
@@ -228,6 +230,7 @@ class AsyncCogniteClientMock(MagicMock, metaclass=_SpecSetEnforcer):
         dm_views = create_autospec(ViewsAPI, instance=True, spec_set=True)
         dm_instances = create_autospec(InstancesAPI, instance=True, spec_set=True)
         dm_graphql = create_autospec(DataModelingGraphQLAPI, instance=True, spec_set=True)
+        dm_records = create_autospec(RecordsAPI, instance=True, spec_set=True)
         dm_streams = create_autospec(StreamsAPI, instance=True, spec_set=True)
         dm_files = create_autospec(DataModelingFilesAPI, instance=True, spec_set=True)
         self.data_modeling = create_autospec(
@@ -240,6 +243,7 @@ class AsyncCogniteClientMock(MagicMock, metaclass=_SpecSetEnforcer):
             instances=dm_instances,
             graphql=dm_graphql,
             statistics=dm_statistics,
+            records=dm_records,
             streams=dm_streams,
             files=dm_files,
         )
@@ -431,6 +435,7 @@ class CogniteClientMock(MagicMock, metaclass=_SpecSetEnforcer):
         dm_views = create_autospec(SyncViewsAPI, instance=True, spec_set=True)
         dm_instances = create_autospec(SyncInstancesAPI, instance=True, spec_set=True)
         dm_graphql = create_autospec(SyncDataModelingGraphQLAPI, instance=True, spec_set=True)
+        dm_records = create_autospec(SyncRecordsAPI, instance=True, spec_set=True)
         dm_streams = create_autospec(SyncStreamsAPI, instance=True, spec_set=True)
         dm_files = create_autospec(SyncDataModelingFilesAPI, instance=True, spec_set=True)
         self.data_modeling = create_autospec(
@@ -443,6 +448,7 @@ class CogniteClientMock(MagicMock, metaclass=_SpecSetEnforcer):
             instances=dm_instances,
             graphql=dm_graphql,
             statistics=dm_statistics,
+            records=dm_records,
             streams=dm_streams,
             files=dm_files,
         )