+++ CHANGELOG.md
@@ -13,6 +13,24 @@ As of 2025-08-29, changes are grouped as follows
 - 🐛 Bug Fixes: Bug fixes.
 - ⚡ Improvements: Transparent changes, e.g. better performance.
 
+## [8.7.0](https://github.com/cognitedata/cognite-sdk-python/compare/cognite-sdk-python-v8.6.0...cognite-sdk-python-v8.7.0) (2026-05-27)
+
+
+### Features
+
+* **Data Modeling:** support usedFor=record on containers ([#2621](https://github.com/cognitedata/cognite-sdk-python/issues/2621)) ([8a8a106](https://github.com/cognitedata/cognite-sdk-python/commit/8a8a106b47f775a78652eee5513edf8398d66f04))
+
+
+### Bug Fixes
+
+* **Files:** Fix pyodide fstat not returning actual file size but 0 ([#2649](https://github.com/cognitedata/cognite-sdk-python/issues/2649)) ([b191b2b](https://github.com/cognitedata/cognite-sdk-python/commit/b191b2b0c615a37c65f1615b115c11ffd3bbccf2))
+* **Functions:** skip pip validation on browser use ([#2638](https://github.com/cognitedata/cognite-sdk-python/issues/2638)) ([c3fbdfa](https://github.com/cognitedata/cognite-sdk-python/commit/c3fbdfaea7ca66d80bc1357b3982b9d35d233427))
+
+
+### Documentation
+
+* **Functions:** clarify type annotation restrictions when using function_handle ([#2639](https://github.com/cognitedata/cognite-sdk-python/issues/2639)) ([98154b0](https://github.com/cognitedata/cognite-sdk-python/commit/98154b01973f853994169d29fad73a52cc6cbd85))
+
 ## [8.6.0](https://github.com/cognitedata/cognite-sdk-python/compare/cognite-sdk-python-v8.5.0...cognite-sdk-python-v8.6.0) (2026-05-20)